backend/requirements.txt
azure-functions>=1.15.0
fastapi>=0.104.0
pydantic>=2.4.0
openai>=1.3.0
httpx>=0.25.0